added delete confirmation alert

+++ app/javascript/controllers/edit_deliverable_controller.js
@@ -7,6 +7,8 @@ import { csrfToken } from "@rails/ujs";
 
 import { Controller } from "stimulus";
 
+import Swal from 'sweetalert2'
+
 export default class extends Controller {
   static targets = ["delete", "details", "edit", "form", "tab"];
   static values = {
@@ -20,16 +22,39 @@ export default class extends Controller {
 
   delete() {
     console.log("deleting");
-    fetch(this.deleteTarget.dataset.href, {
-      method: 'DELETE',
-      headers: { 'Accept': 'text/plain', 'X-CSRF-Token': csrfToken()  } })
-      .then(response => response.text())
-      .then((data) => {
-        console.log(data);
-        this.detailsTarget.innerHTML = ""
-        this.detailsTarget.classList.toggle("d-none")
-      })
+
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#94B3FD',
+      cancelButtonColor: '#bdbdbd',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire(
+          'Deleted!',
+          'Your file has been deleted.',
+          'success'
+        )
+
+        fetch(this.deleteTarget.dataset.href, {
+          method: 'DELETE',
+          headers: { 'Accept': 'text/plain', 'X-CSRF-Token': csrfToken()  } })
+          .then(response => response.text())
+          .then((data) => {
+            console.log(data);
+            this.detailsTarget.innerHTML = ""
+            this.detailsTarget.classList.toggle("d-none")
+          })
+
+      }
+    })
+
   }
+
+
 
   edit() {
     this.tabTarget.classList.toggle("d-none")
